fix(api): return decimals and counts consistently across drivers
PDO rend les `decimal` en chaîne sur MySQL et en nombre sur SQLite. Les
modèles du domaine s'en protègent par un cast `decimal:2` — `Notes`,
`Moyennes`, `PaiementEleve`, `Payment`, `Bourse` — mais `Coefficients`,
`CoefficientMatiere` et `Depense` y avaient échappé : l'API livrait
« 3.00 » en production et 3 en test.

Même écart sur `MessageController::getConversations`, où `non_lus` vient
d'un `SUM(...)` : « 1 » d'un côté, 1 de l'autre. Un compte est un entier,
il est casté comme tel.

`EleveDashboardController` expose `coeff` à côté d'un repli valant `1` :
le champ est numérique, il l'est maintenant sur les deux moteurs.

`MessagesContractTest` interrogeait par ailleurs une colonne `lu` que
`align_notifications_with_frontend` avait supprimée au profit de `read_at`.
Sur SQLite, un identifiant inconnu entre guillemets devient une chaîne
littérale : la requête rendait « lu », `(int) 'lu'` valait 0, et
l'assertion passait sans rien vérifier. MySQL rejette la colonne.

`Coefficients` et `CoefficientMatiere` portent la même table
`coefficient_matieres` — duplication à résorber, signalée en commentaire.

+++ Ecole/Ecole_backend/app/Http/Controllers/Dashboard/EleveDashboardController.php
@@ -44,7 +44,11 @@ class EleveDashboardController extends Controller
             return [
                 'name' => $nom,
                 'note' => round($group->avg('note'), 2),
-                'coeff' => $coefficients[$group->first()->matiere_id] ?? 1,
+                // `(float)` : `coefficient` est un `decimal`, que MySQL rend en
+                // chaîne et SQLite en nombre. Le repli voisin vaut `1`, un
+                // entier — le champ est numérique, il doit l'être sur les deux
+                // moteurs (cf. audit P2.1).
+                'coeff' => (float) ($coefficients[$group->first()->matiere_id] ?? 1),
             ];
         })->values();
